[refactor](be) Align new parquet reader layout
### What problem does this PR solve?

Issue Number: close #xxx

Related PR: #xxx

Problem Summary: Align the new parquet reader file layout with DuckDB-style module boundaries by consolidating reader, scan, and statistics/pruning modules.

### Release note

None

### Check List (For Author)

- Test: Manual test
    - Ran build-support/clang-format.sh
    - Ran git diff --check
    - Fedora DEBUG BE build pending
- Behavior changed: No
- Does this need documentation: No
