--- be/src/format/new_parquet/parquet_scan_planner.h
@@ -1,61 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-#pragma once
-
-#include <memory>
-#include <vector>
-
-#include "common/status.h"
-#include "format/new_parquet/parquet_pruning.h"
-#include "format/new_parquet/selection_vector.h"
-#include "format/reader/file_reader.h"
-
-namespace parquet {
-class FileMetaData;
-class ParquetFileReader;
-} // namespace parquet
-
-namespace doris::parquet {
-
-struct ParquetColumnSchema;
-
-struct ParquetScanRange {
-    int64_t start_offset = 0;
-    int64_t size = -1;
-    int64_t file_size = -1;
-};
-
-struct RowGroupReadPlan {
-    int row_group_id = -1;
-    int64_t first_file_row = 0;
-    int64_t row_group_rows = 0;
-    std::vector<RowRange> selected_ranges;
-};
-
-struct RowGroupScanPlan {
-    std::vector<RowGroupReadPlan> row_groups;
-    ParquetPruningStats pruning_stats;
-};
-
-Status plan_parquet_row_groups(const ::parquet::FileMetaData& metadata,
-                               ::parquet::ParquetFileReader* file_reader,
-                               const std::vector<std::unique_ptr<ParquetColumnSchema>>& file_schema,
-                               const reader::FileScanRequest& request,
-                               const ParquetScanRange& scan_range, RowGroupScanPlan* plan);
-
-} // namespace doris::parquet
