+++ be/src/format/new_parquet/reader/arrow_leaf_reader_adapter.cpp
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "format/new_parquet/column_reader/arrow_leaf_reader_adapter.h"
+#include "format/new_parquet/reader/arrow_leaf_reader_adapter.h"
 
 #include <arrow/array/array_binary.h>
 #include <parquet/api/schema.h>
@@ -29,6 +29,7 @@
 
 #include "core/data_type_serde/decoded_column_view.h"
 #include "core/string_ref.h"
+#include "format/new_parquet/reader/nested_column_reader.h"
 
 namespace doris::parquet {
 namespace {
