+++ be/src/format/new_parquet/reader/list_column_reader.h
@@ -21,27 +21,23 @@
 #include <memory>
 #include <string>
 #include <utility>
-#include <vector>
 
-#include "format/new_parquet/column_reader/column_reader.h"
 #include "format/new_parquet/parquet_column_schema.h"
+#include "format/new_parquet/reader/column_reader.h"
+#include "format/new_parquet/reader/nested_column_reader.h"
 
 namespace doris::parquet {
 
-class StructColumnReader final : public ParquetColumnReader {
+class ListColumnReader final : public ParquetColumnReader {
 public:
-    StructColumnReader(const ParquetColumnSchema& schema, DataTypePtr type,
-                       std::vector<std::unique_ptr<ParquetColumnReader>> children,
-                       std::vector<int> child_output_indices)
+    ListColumnReader(const ParquetColumnSchema& schema, DataTypePtr type,
+                     std::unique_ptr<ParquetColumnReader> element_reader)
             : _field_id(schema.top_level_field_id),
               _nullable_definition_level(schema.nullable_definition_level),
               _repeated_repetition_level(schema.repeated_repetition_level),
               _type(std::move(type)),
               _name(schema.name),
-              _children(std::move(children)),
-              _child_output_indices(std::move(child_output_indices)) {
-        DCHECK_EQ(_children.size(), _child_output_indices.size());
-    }
+              _element_reader(std::move(element_reader)) {}
 
     int file_column_id() const override { return _field_id; }
     int parquet_leaf_column_id() const override { return -1; }
@@ -51,20 +47,18 @@ public:
     Status read(int64_t rows, MutableColumnPtr& column, int64_t* rows_read) override;
     Status skip(int64_t rows) override;
 
-    Status skip_non_scalar_children(int64_t rows);
-    size_t child_count() const { return _children.size(); }
-    ParquetColumnReader* child_reader(size_t child_idx) const { return _children[child_idx].get(); }
-    int child_output_index(size_t child_idx) const { return _child_output_indices[child_idx]; }
     int16_t nullable_definition_level() const { return _nullable_definition_level; }
+    int16_t repeated_repetition_level() const { return _repeated_repetition_level; }
+    ParquetColumnReader* element_reader() const { return _element_reader.get(); }
 
-private:
     int _field_id = -1;
     int16_t _nullable_definition_level = 0;
     int16_t _repeated_repetition_level = 0;
     DataTypePtr _type;
     std::string _name;
-    std::vector<std::unique_ptr<ParquetColumnReader>> _children;
-    std::vector<int> _child_output_indices;
+    std::unique_ptr<ParquetColumnReader> _element_reader;
+    NestedScalarOverflow _element_overflow;
+    NestedStructOverflow _struct_element_overflow;
 };
 
 } // namespace doris::parquet
