+++ be/test/format/new_parquet/parquet_reader_test.cpp
@@ -42,9 +42,9 @@
 #include "core/field.h"
 #include "exprs/vexpr.h"
 #include "exprs/vexpr_context.h"
-#include "format/new_parquet/column_reader/column_reader.h"
 #include "format/new_parquet/parquet_column_schema.h"
-#include "format/new_parquet/parquet_scan_planner.h"
+#include "format/new_parquet/parquet_scan.h"
+#include "format/new_parquet/reader/column_reader.h"
 #include "format/reader/column_mapper.h"
 #include "format/reader/expr/delete_predicate.h"
 #include "format/reader/expr/slot_ref.h"
